Remove thin chat turn execution wrapper
